% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baerhund.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string of the name of the file to be read}
}
\value{
This function returns a tibble of the data file
}
\description{
fars_read
}
\examples{
\dontrun{
fars_read("accident_2015.csv.bz2")
}
}
