% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baerhund.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A string of length 4 which gives the year(s),in the fars filename with month and year attributes.}
}
\value{
returns a list of the summarized data for one or more years
}
\description{
fars_summarize_years
}
\examples{
\dontrun{
fars_summarize_years(years = c(2013, 2014,2015))
}
}
