% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baerhund.R
\name{make_filename}
\alias{make_filename}
\title{baerhund}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string of lenght four which gives the year added to the filename}
}
\value{
This function returns a string with the filename customized with given year
}
\description{
baerhund
}
\examples{
\dontrun{
make_filename(2015)
}
}
