% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baerhund.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A string of length 4 which gives the year(s),in the fars filename with month and year attributes.}
}
\value{
returns a list with the month and year columns from the file
}
\description{
Read data
}
\examples{
\dontrun{fars_read_years(2013:2015)}
}
